% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ine_data.R
\name{get_ine_data}
\alias{get_ine_data}
\title{Get data from the INE API}
\usage{
get_ine_data(
  indicator,
  lang = "PT",
  expected.duration = FALSE,
  max_cells = 30000,
  ...
)
}
\arguments{
\item{indicator}{INE indicator ID as a 7 character string. Example: "0010003".}

\item{lang}{Only "PT" implemented.}

\item{expected.duration}{If TRUE, prints the expected remaining duration in the console.}

\item{max_cells}{Integer smaller than or equal to 40000, the maximum number of cells allowed in each API call.
Default value of 30000}

\item{...}{Values for each Dimension of the indicator.
Each parameter should be in the form dimN, with N one of \{1, ..., Total number of dimensions\}.
If one of the dimensions is not included, output includes all values}
}
\value{
The desired values
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} \cr
A short description...
}
\details{
Logs the expected remaining duration in the console. Can be very high for indicators with many dimensions and/or many unique values.
(E.g. Indicator "0008206 - Deaths by Place of residence, Sex, Age group and Death cause, Annual" will take many hours to extract completely)
}
\examples{
get_ine_data("0010003") # A simple example
get_ine_data("0008206", dim1 = "S7A1996", dim2 = c("11","111"),
             dim4 = c(1,19), dim5 = "TLES") # A more complex example
}
