% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Get indicator metadata}
\usage{
get_metadata(indicator, lang = "PT")
}
\arguments{
\item{indicator}{INE indicator ID as a 7 character string. Example: "0010003".}

\item{lang}{One of "PT" or "EN". Default is "PT".}
}
\value{
API response body as list.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} \cr
Get metadata for the selected indicator from the INE API

Be aware that this function will still return a value and not throw an error if the indicator does not exist.
}
\examples{
get_metadata("0011823")

}
\seealso{
\code{\link[=is_indicator_valid]{is_indicator_valid()}} can be used to check if the indicator is valid before calling this function.
}
