% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_metadata_raw}
\alias{get_metadata_raw}
\title{Get raw indicator metadata}
\usage{
get_metadata_raw(indicator, lang = "PT")
}
\arguments{
\item{indicator}{INE indicator ID as a 7 character string. Example: "0010003".}

\item{lang}{Only "PT" implemented.}
}
\value{
API response body as list.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
\code{get_metadata_raw()} calls the metadata API.

Be aware that this function will still return a value and not throw an error if the indicator does not exist.
}
\seealso{
\code{\link[=is_indicator_valid]{is_indicator_valid()}} can be used to check if the indicator is valid before calling this function.
}
\keyword{internal}
