% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineptR-package.R
\docType{package}
\name{ineptR-package}
\alias{ineptR}
\alias{ineptR-package}
\title{ineptR: Wrapper for Statistics Portugal API}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Set of wrapper and helper functions to facilitate interaction with the Statistics Portugal (Instituto Nacional de Estatistica - INE) API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://c-matos.github.io/ineptR/}
  \item \url{https://github.com/c-matos/ineptR}
}

}
\author{
\strong{Maintainer}: Carlos Matos \email{carlosmdmatos@gmail.com} (\href{https://orcid.org/0000-0003-1134-0396}{ORCID}) [copyright holder]

}
\keyword{internal}
