% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineptR-package.R
\docType{package}
\name{ineptR-package}
\alias{ineptR}
\alias{ineptR-package}
\title{ineptR: Get data from Statistics Portugal (Instituto Nacional de Estatistica - Portugal) to R}
\description{
Set of wrapper and helper functions to facilitate interaction with the Statistics Portugal (Instituto Nacional de Estatistica - INE) API.
}
\author{
\strong{Maintainer}: Carlos Matos \email{first.last@example.com} (\href{https://orcid.org/0000-0003-1134-0396}{ORCID})

}
\keyword{internal}
